Revert changes

+++ src/SimpleSmtp_SSL_QT5/smtp_attachements/smtp.h
@@ -31,8 +31,8 @@ class Smtp : public QObject
 
 
 public:
-	Smtp(const QString &user, const QString &pass,
-		  const QString &host, quint16 port = 465, int timeout = 30000 );
+    Smtp( const QString &user, const QString &pass,
+          const QString &host, int port = 465, int timeout = 30000 );
     ~Smtp();
 
     void sendMail( const QString &from, const QString &to,
@@ -62,7 +62,7 @@ private:
     QString pass;
     QString host;
     static QString status;
-	quint16 port;
+    int port;
     enum states{Tls, HandShake ,Auth,User,Pass,Rcpt,Mail,Data,Init,Body,Quit,Close};
     int state;
